resolver: rename esplora into esplora_blocking and do a dedicated mod
