+++ .github/workflows/build.yml
@@ -44,7 +44,7 @@ jobs:
       fail-fast: false
       matrix:
         feature:
-          - esplora
+          - esplora_blocking
           - log
           - serde
     steps:
